Phase 2: Questcounter/Questcounterlist/Questcounters byte-exact; Players, Questengine, Mapcontrol in progress
Swarm batch A. Whole-tree score 326/337 -> 407/428.

Byte-exact:
- Questcounter / Questcounterlist / Questcounters (19/19). Ghidra's hint names
  were rotated one class off; the module->class mapping recovered from the
  vector element strides is QuestCounter (0x2c, holds vector<QuestCounterList>,
  element stride 8), QuestCounterList (8), QuestCounters (0x20, holds
  vector<QuestCounter>). Emits the previously-missing RTTI element types
  std::vector<QuestCounter> and std::vector<QuestCounterList>.

In progress:
- Players 21/25. New Players.h: vector<Player*> at +0, Player *by_id[100000]
  at +0x20, back-references and counters to +0x61ab0; sizeof 0x61ab4, confirmed
  by `push 0x61ab4` at 0x401afc. Unconverged: Players_Add, CharName_Validate
  (+ the unimplemented Players_Tick/Players_Remove/Character_BuildSaveQuery).
- Questengine 14/19. Layout sizeof 0x94 incl. vector<Quest*>, two
  vector<QuestType> keyword tables. Unconverged: LoadQuest, ParseToken,
  AppendEncoded, GetActionData, GetActionData2.
- Mapcontrol 24/28. Layout sizeof 0x44 (maps vector + start/rescue coords +
  settings/scratch). Unconverged: ToggleDoor, Pub_DecodeNumber_Map,
  AppendEncoded.

Cross-unit fix: the MapObject/MapWarp/MapChest/MapContainer constructors take
int, not short, parameters -- the reference pushes the caller's 32-bit stack
slots directly (0x47cfc2). All four owning units stay byte-exact; this
unblocked Mapcontrol_AddTileSpec/AddWarp/AddLockKey.

Tooling: compare_asm.py canonicalizes `xchg` operand order (symmetric in-
struction; bcc32 and objdump print it either way), which clears the
Players_ActiveCount false negative. Added tests/shortparam_test.cpp probe.

+++ src/Map.cpp
@@ -5,7 +5,7 @@
 
 #pragma package(smart_init)
 
-MapContainer::MapContainer(short map_id, int width, int height)
+MapContainer::MapContainer(int map_id, int width, int height)
 {
     rid = map_id;
     this->width = width;
