+++ src/Mapwarp.cpp
@@ -5,8 +5,7 @@
 
 #pragma package(smart_init)
 
-MapWarp::MapWarp(
-    short from_x, short from_y, int dest_map, int level, short to_x, short to_y)
+MapWarp::MapWarp(int from_x, int from_y, int dest_map, int level, int to_x, int to_y)
 {
     this->from_x = from_x;
     this->from_y = from_y;
